+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-04-27
+
 ### Added
 
 - Windows installer for easy setup of the native messaging host (sets up Chrome, Chrome Beta,
@@ -31,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Logging.
 
-[unreleased]: https://github.com/Tatsh/open-in-mpv/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/Tatsh/open-in-mpv/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/Tatsh/open-in-mpv/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/Tatsh/open-in-mpv/compare/v0.1.2...v0.1.3